New file structure closer to Grin
